+++ core/memory.cpp
@@ -1,4 +1,4 @@
 #include "core/memory.hpp"
 
-u8 memory[SYSTEM::MEMORY];
+u32 memory[SYSTEM::MEMORY];
 u32 stacker = SYSTEM::MEMORY;
