+++ core/memory.hpp
@@ -2,5 +2,5 @@
 
 #include "core/constant.hpp" // IWYU pragma: keep
 
-extern u8 memory[SYSTEM::MEMORY];
+extern u32 memory[SYSTEM::MEMORY];
 extern u32 stacker;
